removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@5338 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/guisim/GUIEdge.cpp
@@ -60,7 +60,7 @@ using namespace std;
 // ===========================================================================
 // included modules
 // ===========================================================================
-GUIEdge::GUIEdge(const std::string &id, size_t numericalID,
+GUIEdge::GUIEdge(const std::string &id, unsigned int numericalID,
                  GUIGlObjectStorage &idStorage) throw()
         : MSEdge(id, numericalID),
         GUIGlObject(idStorage, "edge:" + id)
